pyproject.toml
[project]
name = "cavela-audit"
version = "0.1.0"
description = "Cavela product feedback audit report generator"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "anthropic>=0.40",
    "resend>=2.0",
    "weasyprint>=62.0",
    "markdown-it-py>=3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
